pyproject.toml
[project]
name = "chatbot-ollama"
version = "0.1.0"
description = ""
authors = [
    {name = "Nikhil Pagote",email = "nikhil.pagote@gmail.com"}
]
readme = "README.md"
requires-python = ">=3.12,<4.0"
dependencies = [
    "fastapi (>=0.115.12,<0.116.0)",
    "uvicorn (>=0.34.0,<0.35.0)",
    "requests (>=2.32.3,<3.0.0)",
    "python-multipart",
    "python-dotenv (>=1.1.0,<2.0.0)",
    "jinja2 (>=3.1.6,<4.0.0)",
    "aiofiles>=22.0,<24.0",
    "httpx (>=0.27.0,<1.0.0)",
    "diagrams (>=0.23.4)",
    "langchain",
    "langchain-community",
    "langchain_ollama",
    "langchain_groq",
    "langchain_chroma",
    "langserve",
    "faiss-cpu",
    "pymupdf",
    "langchain-mcp-adapters",
    # Frontend tools (There is dependency conflict if we try to install all of them)
    "streamlit>=1.33.0,<2.0.0",        # Disable Gradio and NiceGUI to use Streamlit.
    #"gradio (==4.44.1)",               # Disable NiceGUI to use Streamlit.
    #"gradio-client (==0.18.1)",        # Gradio client
    #"nicegui>=1.3.17,<2.0.0",         # NiceGUI frontend
    #"reflex>=0.4.7,<1.0.0",            # Reflex (formerly Pynecone)
    
]

[tool.poetry]
packages = [{include = "chatbot_ollama", from = "src"}, { include = "reflex_app", from = "src" }]

[tool.poetry.scripts]
run-reflex = "reflex_app.main:app"

[tool.poetry.group.dev.dependencies]
black = "^25.1.0"
isort = "^6.0.1"

[build-system]
requires = ["poetry-core>=2.0.0,<3.0.0"]
build-backend = "poetry.core.masonry.api"
